Move helper functions to `copyhtml/copyhtml.ts`
Resolves: <https://github.com/EthanThatOneKid/copyhtml/pull/1/files#r1065110680>.
