--- copyhtml/scrape.ts
@@ -1,7 +1,0 @@
-export function scrape(selectors: string[]): string {
-  const elements = selectors.map((selector) =>
-    document.querySelector(selector)
-  );
-  const html = elements.map((element) => element?.outerHTML).join("");
-  return html;
-}
