+++ examples/html/README.md
@@ -16,7 +16,7 @@ configuration.
 Copy and paste the following into your browser's address bar:
 
 ```
-javascript:(()%3D%3E%7Bvar%20p%3D%22Enter%20selectors%20to%20copy%20HTML%20from%20(comma-separated)%3A%22%3Bfunction%20s()%7Blet%20t%3Dprompt(p)%3Bif(t%3D%3D%3Dnull)throw%20new%20Error(%22No%20selectors%20provided.%22)%3Breturn%7Bselectors%3At.split(%22%2C%22).map(e%3D%3Ee.trim())%7D%7Dfunction%20n(t)%7Breturn%20t.map(r%3D%3Edocument.querySelector(r)).map(r%3D%3Er%3F.outerHTML).join(%22%22)%7Dfunction%20c(t)%7Bnavigator.clipboard.writeText(t).catch(o%3D%3E%7Bconsole.error(o)%7D).finally(()%3D%3E%7Bconsole.log(%22Copied%20to%20clipboard.%22%2C%7Btext%3At%7D)%7D)%7Dfunction%20i(t)%7Blet%7Bselectors%3Ao%7D%3Ds()%2Ce%3Dn(o)%3Bc(t%3Ft(e)%3Ae)%7Di()%3B%7D)()%3B%0A
+javascript:(()%3D%3E%7Bfunction%20n(t)%7Blet%20e%3Dc(p())%3Bs(t%3Ft(e)%3Ae)%7Dfunction%20s(t)%7Bnavigator.clipboard.writeText(t).catch(e%3D%3E%7Bconsole.error(e)%7D).finally(()%3D%3E%7Bconsole.log(%22Copied%20to%20clipboard.%22%2C%7Btext%3At%7D)%7D)%7Dfunction%20c(t)%7Breturn%20t.selectors.map(o%3D%3Edocument.querySelector(o)).map(o%3D%3Eo%3F.outerHTML).join(%22%22)%7Dvar%20i%3D%22Enter%20selectors%20to%20copy%20HTML%20from%20(comma-separated)%3A%22%3Bfunction%20p()%7Blet%20t%3Dprompt(i)%3Bif(t%3D%3D%3Dnull)throw%20new%20Error(%22No%20selectors%20provided.%22)%3Breturn%7Bselectors%3At.split(%22%2C%22).map(r%3D%3Er.trim())%7D%7Dn()%3B%7D)()%3B%0A
 ```
 
 > Note: For reusability, you may want to add this to your bookmarks bar.
